Move EntityGroup class and update imports in DatalakeMetadataSettings and JsonMetadataSettings
